Add support for the navigation events
Basically it is possible that the document itself
decides to go to another document location (link to an <a> element
for example). For that we need to cleanup/setup the efl egueb
document interface for both the smart object and the ecore window
In order to simpliofy the code we add another interface for
downloading data. Right now it is based on ecore-con and enesim_stream_file
itself. Later we can add an IO backend too.